create gist feature of post editor add ace code editor for editing gists

+++ grails-app/controllers/codes/recursive/admin/command/PostCommand.groovy
@@ -21,7 +21,7 @@ class PostCommand implements Validateable{
     Date publishedDate = new Date()
     Boolean isPublished = true
     User authoredBy
-    Integer version
+    Long version
     List tags = []
 
     static PostCommand fromPost(Post post) {
